==> notebooks/2.0-train_split.ipynb <==
import sys
sys.path.insert(0, "../")
import numpy as np
import pandas as pd
from src.utils.utils import create_dataset
from pathlib import Path
from tqdm import tqdm
import pandas as pd
import cv2
import matplotlib.pyplot as plt

!ls 

INPUT_PATH = Path("/hdd/kaggle/hubmap/input_v2")
PATH_FOR_CROPS = "/hdd/kaggle/hubmap/input_v2/train_v1_1024"
df_train = pd.read_csv(INPUT_PATH / "train.csv")
train_info = pd.read_csv(INPUT_PATH / "HuBMAP-20-dataset_information.csv")
train_tiffs = list((INPUT_PATH / "train").glob('*.tiff'))

## 5 fold multilabel stratified split

from iterstrat.ml_stratifiers import MultilabelStratifiedKFold

def manual_bin(square):
    if square == 0:
        return 0
    elif 0 < square <=0.1:
        return 1
    elif 0.1 < square <= 0.2:
        return 2
    elif 0.2 < square <= 0.3:
        return 3
    elif 0.3 < square <= 0.4:
        return 4
    elif 0.4 < square <= 0.6:
        return 5
    elif 0.6 < square <= 0.8:
        return 6
    elif 0.8 < square:
        return 7
    
# df['square'] = df['square']/512**2

!ls ../../input/train_1024_256

df = pd.read_csv("../../input/train_1024_256/meta.csv")

df.columns

bining_cols = ['glomerulus_pix', 'medulla', 'cortex', 'outer_stripe',
       'Inner medulla', 'Outer Medulla']
for col in bining_cols:
    df[col] = df[col]/256**2
    df[col] = df[col].apply(manual_bin)

df.head()

df["glomerulus_pix"].value_counts()

df['fold'] = 0

mskf = MultilabelStratifiedKFold(n_splits=5, shuffle=True, random_state=42)
for fold_n, (train_index, test_index) in enumerate(mskf.split(df, df.iloc[:, :7])):
    df.loc[test_index, 'fold'] = fold_n

df.to_csv("/hdd/kaggle/hubmap/input_v2/train_v1_1024/split_v1.csv")

## Group KFold split

df["fold"] = 0
ids = df['img_id'].unique()
d = df['img_id'].value_counts()
split = [
    [0, 10, 9],
    [1, 7, 14],
    [8, 13, 2],
    [6, 3, 12],
    [4, 5, 11],
]
for fold, s in enumerate(split):
    img_ids = d[s].index.tolist()
    df.loc[df['img_id'].isin(img_ids), "fold"] = fold

for i in range(5):
    print(df[df['fold']==i]['img_id'].unique())

!ls ../../input/train_v3_4096_1024

!ls ../../input

df.to_csv("../../input/train_1024_256/split_v1.csv")

# !rm -rf /hdd/kaggle/hubmap/input_v2/train_v1_1024/

ids = df['img_id'].unique()
d = df['img_id'].value_counts()
d

std_min = 1000
x_min = None
for i in tqdm(range(1000000)):
    x = np.random.choice(np.arange(15), size=(5, 3), replace=False)
    m = []
    for i in range(5):
#         print(d[x[i]].sum())
        m.append(d[x[i]].sum())
    mm = np.std(m)
    if mm < std_min:
        std_min = mm
        x_min = x.copy()

print(x_min)
for i in range(5):
    print(d[x_min[i]].sum())

std_min

## Sampling test

FOLD_IMGS = {
    0: ["4ef6695ce", "0486052bb", "2f6ecfcdf"],
    1: ["c68fe75ea", "095bf7a1f", "aaa6a05cc"],
    2: ["afa5e8098", "1e2425f28", "b2dc8411c"],
    3: ["cb2d976f4", "8242609fa", "54f2eec69"],
    4: ["26dc41664", "b9a3865fc", "e79de561c"],
}
def manual_bin(square):
    if square == 0:
        return 0
    elif 0 < square <=0.1:
        return 1
    elif 0.1 < square <= 0.2:
        return 2
    elif 0.2 < square <= 0.3:
        return 3
    elif 0.3 < square <= 0.4:
        return 4
    elif 0.4 < square <= 0.6:
        return 5
    elif 0.6 < square <= 0.8:
        return 6
    elif 0.8 < square:
        return 7

def create_dataset_annotation(df, img_size):
    bining_cols = ['glomerulus_pix', 'medulla', 'cortex', 'outer_stripe',
           'Inner medulla', 'Outer Medulla']
    for col in bining_cols:
        df[col] = df[col]/img_size**2
        df[col] = df[col].apply(manual_bin)
    return df

df = create_dataset_annotation("../../input/train_1024_256/meta.csv")
df.head()

x = Path("../../input/train_1024_256/meta.csv")
x.parent
