==> notebooks/3.0-inference.ipynb <==
import sys
sys.path.insert(0, "../")

import numpy as np
import matplotlib.pyplot as plt
import cv2
import pandas as pd
from pathlib import Path
import segmentation_models_pytorch as smp
import torch
import torch.nn as nn
import tifffile
from torch.utils.data import DataLoader
from src.loops.loops import validation_full_image, inference, inference_overlap
from src.utils.utils import rle2mask, IMAGE_SIZES, mask2rle
from src.transforms.transform import base_transform, valid_transform
from src.datasets.dataset import ImageDataset, SingleTiffDataset
from src.utils.metrics import dice_numpy
import gc
from src.datasets.zarr_dataset import ZarrValidDataset

!ls -l ../../submission/fold_0_4096to1024_epoch_49_score_0.9339.pth

FACTOR = 1
CROP_SIZE= 1024 * 4 * FACTOR
IMG_TRAIN_SIZE = 1024 * FACTOR
BATCH_SIZE = 24
NUM_WORKERS = 0

df = pd.read_csv("/hdd/kaggle/hubmap/input_v2/train.csv").set_index("id", drop=True)
# df_test = pd.read_csv("/hdd/kaggle/hubmap/input_v2/test.csv").set_index("id", drop=True)
df_sub = pd.read_csv("/hdd/kaggle/hubmap/input_v2/sample_submission.csv").set_index("id", drop=True)
df_crops = pd.read_csv("/hdd/kaggle/hubmap/input_v2/train_v1_1024/split_v2.csv")
df_valid = df_crops[df_crops["fold"] == 0].reset_index(drop=True)
img_ids = df_valid["img_id"].unique()

fold_imgs = {
    0: ["4ef6695ce", "0486052bb", "2f6ecfcdf"],
    1: ["c68fe75ea", "095bf7a1f", "aaa6a05cc"],
    2: ["afa5e8098", "1e2425f28", "b2dc8411c"],
    3: ["cb2d976f4", "8242609fa", "54f2eec69"],
    4: ["26dc41664", "b9a3865fc", "e79de561c"],
}

model = smp.Unet("resnet34", encoder_weights=None).cuda()

for fold in range(5):
    print(f"FOLD {fold}")
    w_path = list(Path(f"../first_launch/fold_{fold}").glob("*pth"))[0]
    model.load_state_dict(torch.load(w_path))
    for img_id in fold_imgs[fold]:
        rle = df.loc[img_id, 'encoding']
        df_img = df_valid[df_valid["img_id"] == img_id].reset_index(drop=True)
        loss_fn = nn.BCEWithLogitsLoss()
        img_ds = SingleTiffDataset(
                tiff_path=f"/hdd/kaggle/hubmap/input_v2/train/{img_id}.tiff",
                transform=valid_transform(CROP_SIZE),
                crop_size=CROP_SIZE,
                step = 512,
        )

        img_loader = DataLoader(
                dataset=img_ds,
                batch_size=BATCH_SIZE,
                shuffle=False,
                num_workers=NUM_WORKERS,
                pin_memory=True,
            )
        mask_pred = inference_overlap(img_loader, model, CROP_SIZE).astype(np.float32)
        mask_true = rle2mask(rle, (img_loader.dataset.w, img_loader.dataset.h))
        dice_full_image = dice_numpy(mask_pred, mask_true)
        print(f"{img_id}: {dice_full_image:.4f}")
        cv2.imwrite(f"oof_predicts/{img_id}.png", mask_pred)
        del mask_pred, mask_true
        gc.collect()
        gc.collect()
    
#     metrics  = validation_full_image(img_loader, model, loss_fn, CROP_SIZE, IMAGE_SIZES[img_id], rle, False)
#     dices.append(metrics['dice_full'])
    # metrics, mask_pred, mask_true = validation_full_image(img_loader, model, loss_fn, CROP_SIZE, IMAGE_SIZES[img_id], rle, True)
#     print(metrics)
# print("DICE MEAN:", {np.mean(dices)})

!ls -l ../../submission/1024_256_score_0.9166.pth

model = smp.Unet("resnet34", encoder_weights=None).cuda()
# w_path = "../../submission/fold_0_zarr_epoch_23_score_0.9407.pth"
# w_path = "../../submission/fold_0_old_loader_epoch_38_score_0.9254.pth"
# w_path = "../../submission/fold_0_zarr_pdf_epoch_34_score_0.9123.pth"
w_path = "../weights/4096_1028_random_zarr_zarr_4096_1024/0/epoch_35_score_0.9311.pth"
model.load_state_dict(torch.load(w_path))

!ls -l ../weights/4096_1028_random_zarr_zarr_4096_1024/0/epoch_35_score_0.9311.pth

!ls /hdd/kaggle/hubmap/input_v2/test

fold = 0
print(f"FOLD {fold}")
for img_id in fold_imgs[fold]:
    rle = df.loc[img_id, 'encoding']
    df_img = df_valid[df_valid["img_id"] == img_id].reset_index(drop=True)
    loss_fn = nn.BCEWithLogitsLoss()
    img_ds = SingleTiffDataset(
            tiff_path=f"/hdd/kaggle/hubmap/input_v2/train/{img_id}.tiff",
            transform=valid_transform(IMG_TRAIN_SIZE),
            crop_size=CROP_SIZE,
            step = CROP_SIZE,
    )

    img_loader = DataLoader(
            dataset=img_ds,
            batch_size=BATCH_SIZE,
            shuffle=False,
            num_workers=NUM_WORKERS,
            pin_memory=True,
        )
    mask_pred = inference(img_loader, model, CROP_SIZE, IMG_TRAIN_SIZE).astype(np.float32)
#     input_path = "../../input/zarr_train"
#     val_ds = ZarrValidDataset(
#         img_id,
#         img_path=input_path,
#         transform=valid_transform(IMG_TRAIN_SIZE),
#         crop_size=IMG_TRAIN_SIZE,
#         step=IMG_TRAIN_SIZE,
#     )

#     img_loader = DataLoader(
#         dataset=val_ds,
#         batch_size=BATCH_SIZE,
#         shuffle=False,
#         num_workers=10,
#         pin_memory=True,
#     )

#     metrics_val, mask_pred = validation_full_image(
#         img_loader, model, torch.nn.BCEWithLogitsLoss(), rle=rle, return_mask=True,
#     )
#     print(f"dice_pos: {metrics_val['dice_pos'].4f}, dice_neg: {metrics_val['dice_neg']:.4f}")

    mask_true = rle2mask(rle, (img_loader.dataset.w, img_loader.dataset.h))
    dice_full_image = dice_numpy(mask_pred, mask_true)
    print(f"{img_id}: {dice_full_image:.4f}")
#     cv2.imwrite(f"oof_predicts/{img_id}.png", mask_pred)
#     break
    del mask_pred, mask_true
    gc.collect()
    gc.collect()
    

mask_pred = (mask_pred>0.5).astype(np.uint8)
rle_pred = mask2rle(mask_pred)

pd.DataFrame({"img_id": "57512b7f1", "encoding": rle_pred}, index=[0], columns=['img_id', 'encoding']).to_csv("no_resize_pred.csv")

../weights/zarr_full_image_val/fold_0/fold_0_avg_0.9463.pth
              x1    x2     x4
4ef6695ce: 0.9396 0.9404 0.9411
0486052bb: 0.9496 0.9511 0.9504
2f6ecfcdf: 0.9497 0.9531 0.9520

../weights/zarr_full_image_val/fold_0/epoch_28_score_0.9450.pth
              x1    x4
4ef6695ce: 0.9370 0.9396
0486052bb: 0.9489 0.9509
2f6ecfcdf: 0.9491 0.9507

../weights/zarr_full_image_val/fold_0/epoch_38_score_0.9461.pth
#zarr new    x1     x4
4ef6695ce: 0.9411 0.9425
0486052bb: 0.9473 0.9511
2f6ecfcdf: 0.9499 0.9509

print((0.9398+0.9395+0.9474)/3) #base
print((0.9352+0.9421+0.9494)/3) #zarr dl
print((0.9404+0.9394+0.9520)/3) #avg

np.mean([0.9352, 0.9421, 0.9494])

np.mean([0.9398, 0.9395, 0.9474])

            1024   1024-overlap   2048->512   4096->1024       (4096->1024)x4   (4096->1024)x2  (4096->1024)x6
4ef6695ce: 0.9350    0.9384        0.9359      0.9398              0.9418          0.9413         0.9418
0486052bb: 0.9406    0.9463        0.9299      0.9395              0.9446          0.9405         0.9438
2f6ecfcdf: 0.9485    0.9515        0.9466      0.9474              0.9520          0.9551         0.9508
    
c68fe75ea: 0.7840    0.7886
095bf7a1f: 0.9340    0.9351
aaa6a05cc: 0.9115    0.9229
    
afa5e8098: 0.8858    0.8936
1e2425f28: 0.9330    0.9355
b2dc8411c: 0.9475    0.9475

cb2d976f4: 0.9418    0.9447
8242609fa: 0.9527    0.9536
54f2eec69: 0.9243    0.9273

26dc41664: 0.9405    0.9427
b9a3865fc: 0.9433    0.9448
e79de561c: 0.9241    0.9287

!ls -l


for img_id in df_sub.index:
    img_ds = SingleTiffDataset(
            tiff_path=f"/hdd/kaggle/hubmap/input_v2/test/{img_id}.tiff",
            transform=valid_transform(IMG_TRAIN_SIZE),
            crop_size=CROP_SIZE,
            step = CROP_SIZE,
    )

    img_loader = DataLoader(
            dataset=img_ds,
            batch_size=BATCH_SIZE,
            shuffle=False,
            num_workers=NUM_WORKERS,
            pin_memory=True,
        )
    mask_pred = inference(img_loader, model, CROP_SIZE, IMG_TRAIN_SIZE).astype(np.float32)
    mask_pred = (mask_pred > 0.5).astype(np.uint8)
    mask_rle = mask2rle(mask_pred)
    df_sub.loc[img_id, "predicted"] = mask_rle
#     break
    del mask_pred
    gc.collect()
    gc.collect()
    

df_sub.to_csv("1024_resize.csv")

df_sub

import zarr
z = zarr.zeros((10000, 10000), chunks=(1000, 1000), dtype='i1')

z
