==> notebooks/5.0-deep_eda.ipynb <==
import sys
sys.path.insert(0, "../")

import numpy as np
import matplotlib.pyplot as plt
import cv2
import pandas as pd
from pathlib import Path
import tifffile
from torch.utils.data import DataLoader
from src.loops.loops import validation_full_image, inference, inference_overlap
from src.utils.utils import rle2mask, IMAGE_SIZES, mask2rle
from src.transforms.transform import base_transform, valid_transform
from src.datasets.dataset import ImageDataset, SingleTiffDataset
from src.utils.metrics import dice_numpy
import gc
import rasterio
from tqdm import tqdm

df = pd.read_csv("/hdd/kaggle/hubmap/input_v2/train.csv").set_index("id", drop=True)
df_sub = pd.read_csv("/hdd/kaggle/hubmap/input_v2/sample_submission.csv").set_index("id", drop=True)
df_crops = pd.read_csv("/hdd/kaggle/hubmap/input_v2/train_v1_1024/split_v2.csv")
df_valid = df_crops[df_crops["fold"] == 0].reset_index(drop=True)
img_ids = df_valid["img_id"].unique()

def draw_contours(img, mask, color=(0, 0, 255), thinkness=10):

    contours, _ = cv2.findContours(mask, cv2.RETR_LIST, cv2.CHAIN_APPROX_NONE)
#     print(len(contours))
#     print(type(img), img.dtype)
    for i in range(0, len(contours)):
#         print(i)
        cv2.polylines(img, np.int32(contours[i]), True, color, thinkness)

    
#     return img

# img_id = "57512b7f1"
def get_tiff(img_id, path="/hdd/kaggle/hubmap/input_v2/test"):
    img_path = Path(path) / f"{img_id}.tiff"
    dataset = rasterio.open(img_path, num_threads="all_cpus")
    # mask_main, mask_second = create_masks(img_id, rle, path)

    h, w = dataset.height, dataset.width
    if dataset.count == 3:
        image = dataset.read()#.transpose(1, 2, 0).astype(np.uint8)
        image = np.stack([x for x in image], axis=2).astype(np.uint8)
    else:
        layers = []
        subdatasets = dataset.subdatasets
        if len(subdatasets) > 0:
            for i, subdataset in enumerate(subdatasets, 0):
                layers.append(rasterio.open(subdataset))
        image = np.stack([x.read().squeeze() for x in layers], axis=2).astype(np.uint8)
    return image

df_public = pd.read_csv("submission.csv").set_index("id", drop=True)
df_pred_no_resize = pd.read_csv("1024_no_resize.csv").set_index('id')
df_pred_resize = pd.read_csv("1024_resize.csv").set_index("id", drop=True)

for img_id in df_public.index[1:]:
#     print(img_id)
    image = get_tiff(img_id)
    h, w = image.shape[:2]
    
    mask_no_resize = rle2mask(df_pred_no_resize.loc[img_id, "predicted"], (w, h))
    mask_public = rle2mask(df_public.loc[img_id, "predicted"], (w, h))
    mask_resize = rle2mask(df_pred_resize.loc[img_id, "predicted"], (w, h))

    draw_contours(image, mask_public, (0, 255, 0), thinkness=17)
    draw_contours(image, mask_resize, (255, 0, 0), thinkness=14)
    draw_contours(image, mask_no_resize, (0, 0, 255), thinkness=10)
    cv2.imwrite(f"/hdd/kaggle/hubmap/public_predicted/{img_id}.tiff", image)
    
    del image, mask_no_resize, mask_public, mask_resize
    gc.collect()
    gc.collect()

## Draw contours on train images

for img_id in df.index:
    print(img_id)
    image = get_tiff(img_id, path="/hdd/kaggle/hubmap/input_v2/train")
    h, w = image.shape[:2]
    
    mask_gt = rle2mask(df.loc[img_id, "encoding"], (w, h))
    draw_contours(image, mask_gt, (0, 0, 255), thinkness=10)
    cv2.imwrite(f"/hdd/kaggle/hubmap/train_with_mask/{img_id}.tiff", image)
    
    del image, mask_gt
    gc.collect()
    gc.collect()

!ls /hdd/kaggle/hubmap/train_with_mask/

!ls -l public_predicts/

df_pred_923 = pd.read_csv("public_predicts/924lb.csv").set_index('id')
df_pred_aug = pd.read_csv("public_predicts/1024_512_2_head_897lb.csv").set_index('id')


for img_id in tqdm(df_pred_923.index):
#     print(img_id)
    image = get_tiff(img_id)
    h, w = image.shape[:2]
    
    mask_923 = rle2mask(df_pred_923.loc[img_id, "predicted"], (w, h))
    mask_aug = rle2mask(df_pred_aug.loc[img_id, "predicted"], (w, h))

    draw_contours(image, mask_923, (0, 255, 0), thinkness=10)
    draw_contours(image, mask_aug, (255, 0, 0), thinkness=8)
    cv2.imwrite(f"/hdd/kaggle/hubmap/public_924_and_2048_b3/{img_id}.tiff", image)
    
    del image, mask_923, mask_aug
    gc.collect()
    gc.collect()

!mkdir /hdd/kaggle/hubmap/public_924_and_2048_b3

!ls /hdd/kaggle/hubmap/ -l

!rm -rf /hdd/kaggle/hubmap/public_923_4096 /hdd/kaggle/hubmap/public_923_effnet_0 /hdd/kaggle/hubmap/public_predicted 

!ls ../../input/d48_hand_labelled.csv

df_sub

df_pred_923 = pd.read_csv("public_predicts/v5_r34.csv").set_index('id')
df_pred_aug = pd.read_csv("public_predicts/v5_r34_v2.csv").set_index('id')


for img_id in tqdm(df_sub.index):
#     print(img_id)
    image = get_tiff(img_id)
    h, w = image.shape[:2]
    
    mask_923 = rle2mask(df_pred_923.loc[img_id, "predicted"], (w, h))
    mask_aug = rle2mask(df_pred_aug.loc[img_id, "predicted"], (w, h))

    draw_contours(image, mask_923, (0, 255, 0), thinkness=10)
    draw_contours(image, mask_aug, (255, 0, 0), thinkness=8)
    cv2.imwrite(f"/hdd/kaggle/hubmap/vitaly_timur/{img_id}.tiff", image)
    
    del image, mask_923, mask_aug
    gc.collect()
    gc.collect()

!mkdir /hdd/kaggle/hubmap/vitaly_timur/

!mkdir /hdd/kaggle/hubmap/input_v2/external/labeled
!ls -l /hdd/kaggle/hubmap/input_v2/external/

img_paths = Path("/hdd/kaggle/hubmap/input_v2/external/images_1024").glob("*.png")
masks_paths = Path("/hdd/kaggle/hubmap/input_v2/external/masks_1024").glob("*.png")

img_paths = list(img_paths)

for img_path in tqdm(img_paths):
    im_path = str(img_path)
    mk_path = im_path.replace("images_1024", "masks_1024")
    img = cv2.imread(im_path, -1)
    img = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
    
    mask = cv2.imread(mk_path, -0)
    
    draw_contours(img, mask, (0, 255, 0), thinkness=3)
    cv2.imwrite(f"/hdd/kaggle/hubmap/input_v2/external/labeled/{img_path.name}", img)
    
#     plt.figure(figsize=(7,7))
#     plt.imshow(img)
#     plt.imshow(mask, alpha=0.3)
#     plt.show()
